updated the setup scripts

+++ tools/setup_tmux.sh
@@ -4,13 +4,37 @@
 # Copyright (c) 2017 owl
 #
 
+ROOT=$(cd "$(dirname "$0")" && pwd)
+
 # ----- tmux -----
 # Latest release: https://github.com/tmux/tmux/releases
-curl -L https://github.com/tmux/tmux/releases/download/2.5/tmux-2.5.tar.gz > tmux.tar.gz
-tar vxf tmux.tar.gz
-rm tmux.tar.gz
-cd tmux-2.5
-./configure LDFLAGS='-Wl,-rpath,/usr/local/lib'
-make -j 4
-cd ..
+VER_TMUX_MIN=2.5
+VER_TMUX_NOW=`tmux -V | sed 's/^.* \(.*\)/\1/'`
+#VER_IS_GOOD=$(echo "$VER_TMUX_NOW >= $VER_TMUX_MIN" | bc -l 2>/dev/null)
+
+# Only update tmux if it is neccessary
+if [[ "$VER_IS_GOOD" != "1" ]]; then
+	curl -L https://github.com/tmux/tmux/releases/download/2.5/tmux-2.5.tar.gz > tmux.tar.gz
+	tar vxf tmux.tar.gz
+	rm tmux.tar.gz
+	cd tmux-2.5
+	./configure --prefix="$ROOT/../shell/" CFLAGS="-I$ROOT/../shell/include -I$ROOT/../shell/include/ncurses" LDFLAGS="-L$ROOT/../shell/lib/ -Wl,-rpath,$ROOT/../shell/lib"
+	make -j 4
+	make install
+	cd ..
+	echo "Done; Please continue accordingly"
+else
+	echo "Detected tmux: $VER_TMUX_NOW"
+fi
+
+if ! [ -d tmux-MacOSX* ] ; then
+	git clone https://github.com/ChrisJohnsen/tmux-MacOSX-pasteboard.git
+	cd tmux-MacOSX-pasteboard
+	https://github.com/ChrisJohnsen/tmux-MacOSX-pasteboard.git
+	make -j 2
+	cp reattach-to-user-namespace ../../shell/bin/
+	cd ..
+else
+	echo "Detected tmux MacOSX pasteboard implementation"
+fi
